Concatenating similar components
