+++ components/checkboxes.tsx
@@ -1,11 +1,41 @@
-import Checkbox from "./checkbox"
 import FormGroup from "./form-group"
 import Hint from "./hint"
 import Label from "./label"
 import { FormField, FormOption } from "../lib/types/form"
 
+interface CheckboxProps {
+  index?: number
+  hint?: string
+  label?: string
+  name: string
+  value: string
+}
+
 interface CheckboxesProp {
   field: FormField
+}
+
+export const Checkbox = ({ index, hint, label, name, value }: CheckboxProps): JSX.Element => {
+  let id = name
+  
+  if (index !== undefined) {
+    id += `__${index}`
+  }
+
+  return (
+    <div className="govuk-checkboxes__item">
+      <input
+        className="govuk-checkboxes__input"
+        id={id}
+        name={name}
+        value={value}
+        type="checkbox"
+      />
+      <Label className="govuk-checkboxes__label" content={label || value} htmlFor={id} />
+
+      {hint && <Hint className="govuk-checkboxes__hint" content={hint} />}
+    </div>
+  )
 }
 
 const Checkboxes = ({ field }: CheckboxesProp): JSX.Element => {
